requirements.txt
attrs==20.3.0
bitarray==1.6.1
coverage==5.3
importlib-metadata==3.1.0
iniconfig==1.1.1
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
python-magic==0.4.18
six==1.15.0
toml==0.10.2
zipp==3.4.0
